backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
joblib
scikit-learn
pandas
matplotlib
reportlab